requirements.txt
gitpython==3.1.7
pydriller==1.15.2
pyyaml==5.3.1
options==1.4.10
testresources==2.0.1
dateparser==0.7.6
